+++ lib/src/apps/downloadable_apps/downloadable_apps.dart
@@ -1,5 +1,6 @@
 export 'amap.dart';
 export 'apple_maps.dart';
+export 'capcut.dart';
 export 'facebook.dart';
 export 'google_maps.dart';
 export 'instagram.dart';